requirements.txt
# requirements.txt (Corrected for NumPy 2.0 Compatibility)

# Core Application & CLI
typer[all]==0.12.3
loguru==0.7.2
pydantic==2.8.2
rich==13.7.1

# Local LLM Interface
llama-cpp-python==0.2.79

# Dashboard / Observatory
streamlit==1.36.0

# IPC Communication
pyzmq==25.1.2

# Vector DB & Embeddings (Upgraded)
chromadb==0.5.3
sentence-transformers==2.7.0
onnxruntime==1.18.1